Create part2_main
I should have completed everything before feeding into the LSTM.
You could find our own dictionary in the list called dictionary, and the converted list of utterance lines in the list called converted_list.
I use a smaller file for testing, you could use it if you feel need to.

Sorry but I'm still not familiar enough to write in full python style. In order to save time I brutally wrote part of the code in C style, which caused the increment of functions and  sudden style-changing  in some part. But I think readability is still preserved.

If the Keras part is just like what we discussed last time, the remaining part shouldn't be bothering us that more I hope.